pyproject.toml
[project]
name = "rag-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "chromadb>=1.3.5",
    "faiss-cpu>=1.13.0",
    "ipykernel>=7.1.0",
    "langchain>=1.0.8",
    "langchain-community>=0.4.1",
    "langchain-core>=1.0.6",
    "langchain-google-genai>=3.1.0",
    "langgraph>=1.0.3",
    "pymupdf>=1.26.6",
    "pypdf>=6.3.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.1.2",
    "typesense>=1.3.0",
]
